Add issue to example

+++ examples/example-1/src/lib.rs
@@ -6,7 +6,7 @@ pub struct MyStruct;
 
 #[contractimpl]
 impl MyStruct {
-    pub fn split_profit(percentage: u64, total_profit: u64) -> u64 {
+    pub fn split_profit(percentage: u32, total_profit: u32) -> u32 {
         (percentage / 100) * total_profit
     }
 }
